requirements.txt
# Core (CUDA 11.8 for Py3.9 compatibility)
torch==2.0.1+cu118
torchvision==0.15.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

diffusers==0.19.3
transformers==4.31.0
accelerate==0.21.0
huggingface-hub==0.16.4

fastapi==0.95.2
uvicorn==0.22.0

numpy==1.23.5 
pillow==9.5.0
sse-starlette>=1.6.5

apscheduler==3.10.4
fastapi-pagination==0.14.1

sqlalchemy==2.0.23
pyodbc==4.0.39
python-dotenv==1.0.0
alembic==1.12.1
